LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-compact-controls-bar-styles.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<style type="text/css" media="screen">
    
.media-controls {
    position: absolute;
    top: 0;
    left: 0;
}
    
</style>
<script type="text/javascript">

description("Testing the <code>MacOSInlineMediaControls</code> computed styles in compact mode.");

window.jsTestIsAsync = true;

const mediaControls = new MacOSInlineMediaControls({ width: 240 });
mediaControls.layoutTraits = LayoutTraits.macOS | LayoutTraits.Compact;

let bounds;
scheduler.frameDidFire = function()
{
    document.body.appendChild(mediaControls.element);

    shouldBe("mediaControls.controlsBar.element.getBoundingClientRect().height", "25");

    mediaControls.element.remove();

    debug("");
    finishMediaControlsTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
