LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-compact-controls-bar-styles-expected.txt

Testing the MacOSInlineMediaControls computed styles in compact mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.controlsBar.element.getBoundingClientRect().height is 25

PASS successfullyParsed is true

TEST COMPLETE
